begin migration to AdminUserProfile
temp route to AdminUserProfile
