+++ ui/src/app/pages/admin/admin-user-profile.tsx
@@ -1,0 +1,10 @@
+import React, {useEffect} from 'react';
+
+import {styles} from './admin-user-common';
+import {FadeBox} from 'app/components/containers';
+import {WithSpinnerOverlayProps} from 'app/components/with-spinner-overlay';
+
+export const AdminUserProfile = (spinnerProps: WithSpinnerOverlayProps) => {
+  useEffect(() => spinnerProps.hideSpinner(), []);
+  return <FadeBox style={styles.fadeBox}>temp</FadeBox>;
+}
